# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "cf-teamspeed API",
    "version": "1.0.0",
    "description": "基于 Cloudflare Realtime SFU + Workers 的私人游戏语音室 API。\n\n**这份文档覆盖完整的核心闭环**（客户端最少需要实现的部分）：配置、鉴权与续期、房间生命周期、WebSocket 握手票、SFU 信令、在线状态。\n\n其余接口（房间管理、管理后台）见 `docs/API.md`，那份是完整清单；本文件刻意保持精简，以便作为「新客户端接入检查表」使用。\n\n鉴权：除 `security: []` 的接口外，一律带 `Authorization: Bearer <token>`。\n\n错误：统一返回 `{ok:false, error, code}`；`code` 是稳定契约，分支判断请用它而不是 `error` 文案。",
    "license": { "name": "MIT" }
  },
  "servers": [
    { "url": "https://ts.futurvo.cc", "description": "示例部署（请替换成你自己的 Worker 域名）" }
  ],
...
  "security": [{ "bearerAuth": [] }],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "登录接口返回的 `token`。有效期 12 小时，随请求滚动续期；服务端会通过响应头 `X-Refreshed-Token` 下发新 token。"
      }
    },
...
    "schemas": {
      "ApiOk": {
        "type": "object",
        "required": ["ok", "data"],
        "properties": {
          "ok": { "const": true },
          "data": {}
        }
      },
...
    },
...
  },
  "paths": {
...
    "/api/presence/heartbeat": {
      "post": {
        "tags": ["presence"],
        "summary": "登记在线状态与所在房间",
        "description": "建议 15 秒一次。被移出服务器时返回 `{alive:false, banned:true}`（成功响应 + 标记）—— 高频心跳里抛 403 容易被客户端当成网络抖动忽略掉。看到 `banned:true` 就清本地会话并回登录。",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "roomId": { "type": ["string", "null"] },
                  "roomName": { "type": ["string", "null"] },
                  "status": { "type": "string", "enum": ["online", "busy", "away", "invisible"] },
                  "invitable": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "已在册（或已被移出服务器）",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiOk" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "alive": { "type": "boolean" },
                            "banned": { "type": "boolean", "description": "true = 已被移出服务器，客户端应清会话回登录" }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/api/presence/invite": {
      "post": {
        "tags": ["presence"],
        "summary": "邀请他人进房间",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["toUid", "roomId"],
                "properties": {
                  "toUid": { "type": "string" },
                  "roomId": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "已投递（或对方状态不允许，`reason` 说明原因）" },
          "403": { "description": "你不在这个房间里（`FORBIDDEN`）" }
        }
      }
    },
    "/api/presence/kick": {
      "post": {
        "tags": ["presence"],
        "summary": "踢出服务器（管理员）",
        "description": "需要客户端会话且 `role=admin`：断开在线状态 + 封禁昵称 + 移出所在房间。",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["uid"],
                "properties": {
                  "uid": { "type": "string" },
                  "reason": { "type": "string" },
                  "durationMinutes": { "type": "integer", "description": "不填为永久" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "已踢出并封禁" },
          "403": { "description": "权限不足（`FORBIDDEN`）" }
        }
      }
    }
  },
...
}